sref/kernels: bf16 CSD parity guard + fix pre-existing exp2 UB (deep-sweep extras)
While Phase-3 trained, extra train<->infer robustness checks:
- CSD parity fixture switched to bf16 (the real arm's export format) so make test
  guards the bf16 dequant/load path; 15/15 still exact.
- AddressSanitizer: C perceive/adapter path memory-clean (no overrun/UAF).
- UBSan found a pre-existing (non-CSD) UB in the exp2 kernel (iris_kernels.h:30,
  left-shift of negative) — fixed with a numerically-identical unsigned shift.
  UBSan clean, make test-unit all-green, iris rebuilt (make mps).

+++ debug/gen_ip_adapter_fixture.py
@@ -177,7 +177,7 @@ def main() -> int:
         ckpt = Path(td) / "step_0000001.safetensors"
         mx.save_safetensors(str(ckpt), flat_csd)
         bdir = out / "bundle_csd"
-        _export(ckpt, bdir, "float16")
+        _export(ckpt, bdir, "bfloat16")   # the REAL arm exports bf16 — guard that load path
         _reload_from_bundle(csd_model, bdir, inv=_INV_CSD)
         es, cs = _dump_goldens(csd_model, out, "_csd", csd_vec, img_q_flat)
         print(f"  csd     : ip_embeds std={es:.4f}  contrib std={cs:.4f}")
